+++ cmd/datakit/cmds/checkcfg.go
@@ -2,8 +2,9 @@ package cmds
 
 import (
 	"fmt"
+	"path/filepath"
 	"reflect"
-	"time"
+	"strings"
 
 	"github.com/influxdata/toml"
 	"github.com/influxdata/toml/ast"
@@ -13,10 +14,11 @@ import (
 )
 
 var (
-	failed  = 0
-	unknown = 0
+	failed  = []string{}
+	unknown = []string{}
+	ignored = []string{}
 	passed  = 0
-	ignored = 0
+	checked = 0
 )
 
 func checkInputCfg(tpl *ast.Table, fp string) {
@@ -24,7 +26,7 @@ func checkInputCfg(tpl *ast.Table, fp string) {
 
 	if len(tpl.Fields) == 0 {
 		warnf("[W] no content in %s\n", fp)
-		ignored++
+		ignored = append(ignored, fp)
 		return
 	}
 
@@ -32,7 +34,7 @@ func checkInputCfg(tpl *ast.Table, fp string) {
 		switch field {
 		default:
 			infof("[I] ignore config %s\n", fp)
-			ignored++
+			ignored = append(ignored, fp)
 			return
 
 		case "inputs": //nolint:goconst
@@ -43,11 +45,11 @@ func checkInputCfg(tpl *ast.Table, fp string) {
 				for inputName, v := range stbl.Fields {
 					if c, ok := inputs.Inputs[inputName]; !ok {
 						warnf("[W] unknown input `%s' found in %s\n", inputName, fp)
-						unknown++
+						unknown = append(unknown, fp)
 					} else {
 						if _, err = config.TryUnmarshal(v, inputName, c); err != nil {
 							errorf("[E] failed to init input %s from %s:\n%s\n", inputName, fp, err.Error())
-							failed++
+							failed = append(failed, fp+": "+err.Error())
 						} else {
 							if FlagVVV {
 								output("[OK] %s/%s\n", inputName, fp)
@@ -61,71 +63,95 @@ func checkInputCfg(tpl *ast.Table, fp string) {
 	}
 }
 
+func showCheckResult() {
+	infof("\n------------------------\n")
+	infof("checked %d samples, %d ignored, %d passed, %d failed, %d unknown, ",
+		checked, len(ignored), passed, len(failed), len(unknown))
+
+	if len(ignored) > 0 {
+		infof("ignored:\n")
+		for _, x := range ignored {
+			infof("\t%s\n", x)
+		}
+	}
+
+	if len(unknown) > 0 {
+		infof("unknown:\n")
+		for _, x := range unknown {
+			warnf("\t%s\n", x)
+		}
+	}
+
+	if len(failed) > 0 {
+		infof("failed:\n")
+		for _, x := range failed {
+			errorf("\t%s\n", x)
+		}
+	}
+}
+
 // check samples of every inputs.
 func checkSample() error {
-	start := time.Now()
-	failed = 0
-	unknown = 0
+	failed = []string{}
+	unknown = []string{}
 	passed = 0
-	ignored = 0
+	checked = len(inputs.Inputs)
+	ignored = []string{}
 
 	for k, c := range inputs.Inputs {
 		i := c()
 
 		if k == datakit.DatakitInputName {
 			warnf("[W] ignore self input\n")
-			ignored++
+			ignored = append(ignored, k)
 			continue
 		}
 
 		tpl, err := toml.Parse([]byte(i.SampleConfig()))
 		if err != nil {
 			errorf("[E] failed to parse %s: %s", k, err.Error())
-			failed++
+			failed = append(failed, k+": "+err.Error())
 		} else {
 			checkInputCfg(tpl, k)
 		}
 	}
 
-	infof("\n------------------------\n")
-	infof("checked %d sample, %d ignored, %d passed, %d failed, %d unknown, ",
-		len(inputs.Inputs), ignored, passed, failed, unknown)
+	showCheckResult()
 
-	infof("cost %v\n", time.Since(start))
-
-	if failed > 0 {
-		return fmt.Errorf("load %d sample failed", failed)
+	if len(failed) > 0 {
+		return fmt.Errorf("load %v sample failed", failed)
 	}
 	return nil
 }
 
 func checkConfig(dir, suffix string) error {
-	start := time.Now()
 	fps := config.SearchDir(dir, suffix)
 
-	failed = 0
-	unknown = 0
+	failed = []string{}
+	unknown = []string{}
 	passed = 0
-	ignored = 0
+	checked = 0
+	ignored = []string{}
 
 	for _, fp := range fps {
+		// Skip hidden files.
+		if strings.HasPrefix(filepath.Base(fp), ".") {
+			continue
+		}
 		tpl, err := config.ParseCfgFile(fp)
 		if err != nil {
 			errorf("[E] failed to parse %s: %s, %s", fp, err.Error(), reflect.TypeOf(err))
-			failed++
+			failed = append(failed, fp+": "+err.Error())
 		} else {
 			checkInputCfg(tpl, fp)
 		}
+		checked++
 	}
 
-	infof("\n------------------------\n")
-	infof("checked %d conf, %d ignored, %d passed, %d failed, %d unknown, ",
-		len(fps), ignored, passed, failed, unknown)
+	showCheckResult()
 
-	infof("cost %v\n", time.Since(start))
-
-	if failed > 0 {
-		return fmt.Errorf("load %d conf failed", failed)
+	if len(failed) > 0 {
+		return fmt.Errorf("load %d conf failed", len(failed))
 	}
 
 	return nil