+++ plugins/externals/ebpf/dnsflow/dnsflow.go
@@ -1,5 +1,5 @@
-//go:build (linux && ignore) || ebpf
-// +build linux,ignore ebpf
+//go:build (linux && amd64 && ebpf) || (linux && arm64 && ebpf)
+// +build linux,amd64,ebpf linux,arm64,ebpf
 
 package dnsflow