+++ plugins/inputs/inputs.go
@@ -24,13 +24,12 @@ const (
 )
 
 var (
-	Inputs     = map[string]Creator{}
-	InputsInfo = map[string][]*inputInfo{}
-	ConfigInfo = map[string]*Config{}
-
-	l           = logger.DefaultSLogger("inputs")
+	Inputs      = map[string]Creator{}
+	InputsInfo  = map[string][]*inputInfo{}
+	ConfigInfo  = map[string]*Config{}
 	panicInputs = map[string]int{}
 	mtx         = sync.RWMutex{}
+	l           = logger.DefaultSLogger("inputs")
 )
 
 func GetElectionInputs() []ElectionInput {
@@ -43,6 +42,7 @@ func GetElectionInputs() []ElectionInput {
 			}
 		}
 	}
+
 	return res
 }
 
@@ -224,6 +224,25 @@ func RunInputs(isReload bool) error {
 	return nil
 }
 
+func RunInputExtra() error {
+	mtx.RLock()
+	defer mtx.RUnlock()
+
+	for name, arr := range InputsInfo {
+		for _, ii := range arr {
+			if ii.input == nil {
+				l.Debugf("skip non-datakit-input %s", name)
+				continue
+			}
+
+			if inp, ok := ii.input.(HTTPInput); ok {
+				inp.RegHTTPHandler()
+			}
+		}
+	}
+	return nil
+}
+
 func StopInputs() error {
 	mtx.RLock()
 	defer mtx.RUnlock()
@@ -259,20 +278,20 @@ func protectRunningInput(name string, ii *inputInfo) {
 			crashTime = append(crashTime, fmt.Sprintf("%v", time.Now()))
 			addPanic(name)
 
-			FeedReporter(&io.Reporter{Status: "warning", Message: string(trace), Category: "input"})
+			io.FeedEventLog(&io.Reporter{Status: "warning", Message: string(trace), Category: "input"})
 
 			if len(crashTime) >= MaxCrash {
 				l.Warnf("input %s crash %d times(at %+#v), exit now.",
 					name, len(crashTime), strings.Join(crashTime, "\n"))
 
 				message := fmt.Sprintf("input '%s' has exceeded the max crash times %v and it will be stopped.", name, MaxCrash)
-				FeedReporter(&io.Reporter{Message: message, Status: "error", Category: "input"})
+				io.FeedEventLog(&io.Reporter{Message: message, Status: "error", Category: "input"})
 
 				return
 			}
 		} else {
 			message := fmt.Sprintf("input '%s' starts to run, totally %v instances.", name, InputEnabled(name))
-			FeedReporter(&io.Reporter{Message: message, Category: "input"})
+			io.FeedEventLog(&io.Reporter{Message: message, Category: "input"})
 		}
 
 		select {
@@ -309,7 +328,6 @@ func InputEnabled(name string) (n int) {
 	}
 
 	n = len(arr)
-	l.Debugf("name %s enabled %d", name, n)
 	return
 }