+++ plugins/inputs/redis/input.go
@@ -261,6 +261,7 @@ func (i *Input) RunPipeline() {
 
 func (i *Input) Run() {
 	l = logger.SLogger("redis")
+	io.FeedEventLog(&io.Reporter{Message: inputName + " start ok, ready for collecting metrics.", Logtype: "event"})
 
 	i.Interval.Duration = config.ProtectedInterval(minInterval, maxInterval, i.Interval.Duration)
 
@@ -291,6 +292,7 @@ func (i *Input) Run() {
 		i.collectSlowlogMeasurement,
 		i.collectDBMeasurement,
 		i.CollectLatencyMeasurement,
+		i.collectReplicaMeasurement,
 	}
 
 	// 判断是否采集集群